before swiper
